requirements.txt
python-dotenv
# requirements
telebot==0.0.5
gspread==6.1.2
